chore: Create NotFound Component
